backend/requirements.txt.txt
fastapi
uvicorn
transformers
torch
pytesseract
pdfplumber
pillow
unfluff
requests
